+++ ReverseAD/src/checkpointing/checkpoint_trace.cpp
@@ -10,8 +10,8 @@ namespace ReverseAD {
 
 CheckpointTrace::CheckpointTrace() {
   _checkpoint_num = 0;
-  _tape_x_num = std::make_shared<TrivialTape<int>>();
-  _tape_iter_num = std::make_shared<TrivialTape<int>>();
+  _tape_x_num = std::make_shared<TrivialTape<size_t>>();
+  _tape_iter_num = std::make_shared<TrivialTape<size_t>>();
   _tape_x_values = std::make_shared<VirtualTape<double>>();
   _tape_x_locint = std::make_shared<VirtualTape<locint>>();
   _tape_runtime_env = std::make_shared<TrivialTape<RuntimeEnv>>();
@@ -44,11 +44,11 @@ void CheckpointTrace::end_reverse() {
   _tape_runtime_env->end_reverse();
 }
 
-int CheckpointTrace::make_checkpoint(
-    adouble* x, int x_num, std::shared_ptr<RuntimeEnv>& runtime_env) {
+size_t CheckpointTrace::make_checkpoint(
+    adouble* x, size_t x_num, std::shared_ptr<RuntimeEnv>& runtime_env) {
   _checkpoint_num++; 
   _tape_x_num->put(x_num);
-  for (int i = 0; i < x_num; i++) {
+  for (size_t i = 0; i < x_num; i++) {
     _tape_x_values->put(x[i].getVal());
     _tape_x_locint->put(x[i].getLoc());
   }
@@ -56,20 +56,20 @@ int CheckpointTrace::make_checkpoint(
   return _checkpoint_num;
 }
 
-void CheckpointTrace::set_iteration_num(int iter_num) {
+void CheckpointTrace::set_iteration_num(size_t iter_num) {
   _tape_iter_num->put(iter_num);
 }
-int CheckpointTrace::get_iteration_num() {
+size_t CheckpointTrace::get_iteration_num() {
   return _tape_iter_num->get_next_r();
 }
-int CheckpointTrace::get_checkpoint(
-   adouble* x, int x_num, std::shared_ptr<RuntimeEnv>& runtime_env) {
+size_t CheckpointTrace::get_checkpoint(
+   adouble* x, size_t x_num, std::shared_ptr<RuntimeEnv>& runtime_env) {
   if (_tape_x_num->get_next_r() != x_num) {
     std::cerr << "checkpointing x_num mismatch" << std::endl;
   }
 
   // In place construction
-  for (int i = x_num - 1; i >= 0; i--) {
+  for (size_t i = x_num - 1; i >= 0; i--) {
     new(&x[i]) adouble(_tape_x_values->get_next_r(),
                        _tape_x_locint->get_next_r());
   }
