+++ ReverseAD/src/driver/forward_over_reverse.cpp
@@ -13,12 +13,12 @@ namespace ReverseAD {
 
 std::shared_ptr<TrivialTrace<SingleForward>> get_forward_trace(
     std::shared_ptr<TrivialTrace<double>> trace,
-    int ind_size,
+    size_t ind_size,
     double* ind_init_value,
     double* adjoint_init_value) {
   void* raw_memory = ::operator new[](ind_size * sizeof(SingleForward));
   SingleForward* x = static_cast<SingleForward*>(raw_memory);
-  for(int i=0; i<ind_size; i++) {
+  for(size_t i = 0; i < ind_size; i++) {
     new(&x[i]) SingleForward(ind_init_value[i], adjoint_init_value[i]);
   }
   std::shared_ptr<TrivialTrace<SingleForward>> new_trace =
@@ -29,8 +29,8 @@ std::shared_ptr<TrivialTrace<SingleForward>> get_forward_trace(
 }
 
 void forward_over_reverse(std::shared_ptr<TrivialTrace<double>> trace,
-                          int ind_size,
-                          int dep_size,
+                          size_t ind_size,
+                          size_t dep_size,
                           double* ind_init_value,
                           double* adjoint_init_value,
                           double** hessian_vector) {
@@ -39,25 +39,25 @@ void forward_over_reverse(std::shared_ptr<TrivialTrace<double>> trace,
 
   BaseReverseAdjoint<SingleForward> adjoint(new_trace);
   adjoint.compute(ind_size, dep_size);
-  std::shared_ptr<DerivativeTensor<int, SingleForward>> tensor = adjoint.get_tensor();
+  std::shared_ptr<DerivativeTensor<size_t, SingleForward>> tensor = adjoint.get_tensor();
 
   if (tensor->get_dep_size() != dep_size) {
     // TODO(muwang) some error here.
   }
 
-  int size;
-  int** tind;
+  size_t size;
+  size_t** tind;
   SingleForward* values;
-  for (int i=0; i<dep_size; i++) {
+  for (size_t i = 0; i < dep_size; i++) {
     tensor->get_internal_coordinate_list(i, 1, &size, &tind, &values);
     //std::cout << "size = " << size << std::endl;
     if (size != ind_size) {
       // some error here.
     }
-    for (int j = 0; j < ind_size; j++) {
+    for (size_t j = 0; j < ind_size; j++) {
       hessian_vector[i][j] = 0.0;
     }
-    for (int j = 0; j < size; j++) {
+    for (size_t j = 0; j < size; j++) {
       //std::cout << "tind[j][0] = " << tind[j][0] << std::endl;
       hessian_vector[i][tind[j][0]] = values[j].getDer();
     }
@@ -65,8 +65,8 @@ void forward_over_reverse(std::shared_ptr<TrivialTrace<double>> trace,
 }
 
 void forward_over_second(std::shared_ptr<TrivialTrace<double>> trace,
-                         int ind_size,
-                         int dep_size,
+                         size_t ind_size,
+                         size_t dep_size,
                          double* ind_init_value,
                          double* adjoint_init_value,
                          double*** third_vector) {
@@ -75,24 +75,24 @@ void forward_over_second(std::shared_ptr<TrivialTrace<double>> trace,
 
   BaseReverseHessian<SingleForward> hessian(new_trace);
   hessian.compute(ind_size, dep_size);
-  std::shared_ptr<DerivativeTensor<int, SingleForward>> tensor =
+  std::shared_ptr<DerivativeTensor<size_t, SingleForward>> tensor =
       hessian.get_tensor();
 
   if (tensor->get_dep_size() != dep_size) {
     // TODO(muwang) some error here.
   }
-  int size;
-  int** tind;
+  size_t size;
+  size_t** tind;
   SingleForward* values;
-  for (int i=0; i<dep_size; i++) {
+  for (size_t i=0; i<dep_size; i++) {
     tensor->get_internal_coordinate_list(i, 2, &size, &tind, &values);
     //std::cout << "size = " << size << std::endl;
-    for (int j = 0; j < ind_size; j++) {
-      for(int k = 0; k < ind_size; k++) {
+    for (size_t j = 0; j < ind_size; j++) {
+      for(size_t k = 0; k < ind_size; k++) {
         third_vector[i][j][k] = 0.0;
       }
     }
-    for (int j = 0; j < size; j++) {
+    for (size_t j = 0; j < size; j++) {
       //std::cout << "tind[j][0] = " << tind[j][0] << std::endl;
       //std::cout << "tind[j][1] = " << tind[j][1] << std::endl; 
       third_vector[i][tind[j][0]][tind[j][1]] = values[j].getDer();
@@ -101,8 +101,8 @@ void forward_over_second(std::shared_ptr<TrivialTrace<double>> trace,
 }
 
 void forward_over_third(std::shared_ptr<TrivialTrace<double>> trace,
-                        int ind_size,
-                        int dep_size,
+                        size_t ind_size,
+                        size_t dep_size,
                         double* ind_init_value,
                         double* adjoint_init_value,
                         double**** fourth_vector) {
@@ -111,29 +111,29 @@ void forward_over_third(std::shared_ptr<TrivialTrace<double>> trace,
 
   BaseReverseThird<SingleForward> third(new_trace);
   third.compute(ind_size, dep_size);
-  std::shared_ptr<DerivativeTensor<int, SingleForward>> tensor =
+  std::shared_ptr<DerivativeTensor<size_t, SingleForward>> tensor =
       third.get_tensor();
 
   if (tensor->get_dep_size() != dep_size) {
     // TODO(muwang) some error here.
   }
-  int size;
-  int** tind;
+  size_t size;
+  size_t** tind;
   SingleForward* values;
-  for (int i=0; i<dep_size; i++) {
+  for (size_t i=0; i<dep_size; i++) {
     tensor->get_internal_coordinate_list(i, 3, &size, &tind, &values);
     //std::cout << "size = " << size << std::endl;
     if (size != ind_size) {
       // some error here.
     }
-    for (int j = 0; j < ind_size; j++) {
-      for(int k = 0; k < ind_size; k++) {
-        for (int l = 0; l<ind_size; l++) {
+    for (size_t j = 0; j < ind_size; j++) {
+      for(size_t k = 0; k < ind_size; k++) {
+        for (size_t l = 0; l<ind_size; l++) {
           fourth_vector[i][j][k][l] = 0.0;
         }
       }
     }
-    for (int j = 0; j < size; j++) {
+    for (size_t j = 0; j < size; j++) {
       //std::cout << "tind[j][0] = " << tind[j][0] << std::endl;
       //std::cout << "tind[j][1] = " << tind[j][1] << std::endl; 
       //std::cout << "tind[j][2] = " << tind[j][2] << std::endl; 
@@ -142,22 +142,22 @@ void forward_over_third(std::shared_ptr<TrivialTrace<double>> trace,
   } 
 }
 
-std::shared_ptr<DerivativeTensor<int, double>> strip_derivative(
-    const std::shared_ptr<DerivativeTensor<int, SingleForward>> tensor,
-    int t_order,
-    int ind_size,
-    int dep_size) {
-  std::shared_ptr<DerivativeTensor<int, double>> ret =
-      std::make_shared<DerivativeTensor<int, double>>(dep_size, ind_size, t_order);
-  int size;
-  int** tind;
+std::shared_ptr<DerivativeTensor<size_t, double>> strip_derivative(
+    const std::shared_ptr<DerivativeTensor<size_t, SingleForward>> tensor,
+    size_t t_order,
+    size_t ind_size,
+    size_t dep_size) {
+  std::shared_ptr<DerivativeTensor<size_t, double>> ret =
+      std::make_shared<DerivativeTensor<size_t, double>>(dep_size, ind_size, t_order);
+  size_t size;
+  size_t** tind;
   SingleForward* value;
-  for (int i=0; i<dep_size; i++) {
+  for (size_t i = 0; i < dep_size; i++) {
     ret->put_dep_value(i, tensor->get_dep_value(i).getDer());
-    for (int j=1; j<=t_order; j++) {
+    for (size_t j = 1; j <= t_order; j++) {
       tensor->get_internal_coordinate_list(i, j, &size, &tind, &value);
       ret->init_single_tensor(i, j, size);
-      for (int k=0; k<size; k++) {
+      for (size_t k = 0; k < size; k++) {
         ret->put_value(i, j, k, tind[k], value[k].getDer());
       }
     }
@@ -165,17 +165,17 @@ std::shared_ptr<DerivativeTensor<int, double>> strip_derivative(
   return ret;
 }
 
-std::shared_ptr<DerivativeTensor<int, double>> directional_reverse(
+std::shared_ptr<DerivativeTensor<size_t, double>> directional_reverse(
     std::shared_ptr<TrivialTrace<double>> trace,
-    int t_order,
-    int ind_size,
-    int dep_size,
+    size_t t_order,
+    size_t ind_size,
+    size_t dep_size,
     double* ind_init_value,
     double* adjoint_init_value) {
   std::shared_ptr<TrivialTrace<SingleForward>> new_trace =
       get_forward_trace(trace, ind_size, ind_init_value, adjoint_init_value);
 
-  std::shared_ptr<DerivativeTensor<int, SingleForward>> tensor;
+  std::shared_ptr<DerivativeTensor<size_t, SingleForward>> tensor;
   if (t_order == 1) {
     BaseReverseAdjoint<SingleForward> adjoint(new_trace);
     adjoint.compute(ind_size, dep_size);
