+++ ReverseAD/test/algorithm/test_base_reverse.cpp
@@ -10,27 +10,27 @@ using ReverseAD::TrivialTrace;
 using ReverseAD::DerivativeTensor;
 
 
-void check_answer(std::shared_ptr<DerivativeTensor<int, double>> tensor) {
-  int size;
-  int** tind;
+void check_answer(std::shared_ptr<DerivativeTensor<size_t, double>> tensor) {
+  size_t size;
+  size_t** tind;
   double* values;
 
   tensor->get_internal_coordinate_list(0, 1, &size, &tind, &values);
 
   std::cout << "adjoint size = "<<size << std::endl;
-  for(int i = 0; i < size; i++) {
+  for(size_t i = 0; i < size; i++) {
     std::cout << "A[" << tind[i][0] << "] = " << values[i] << std::endl;
   }
 
   tensor->get_internal_coordinate_list(0, 2, &size, &tind, &values);
   std::cout << "hessian size = " <<size << std::endl;
-  for(int i = 0; i < size; i++) {
+  for(size_t i = 0; i < size; i++) {
     std::cout << "H["<<tind[i][0]<<","<<tind[i][1]<<"] = "<<values[i] << std::endl;
   }
 
   tensor->get_internal_coordinate_list(0, 3, &size, &tind, &values);
   std::cout << "third order size = " << size << std::endl;
-  for (int i=0; i<size; i++) {
+  for (size_t i=0; i<size; i++) {
     std::cout << "T[" << tind[i][0] << ", " << tind[i][1]
               << ", " << tind[i][2] << " ] =" << values[i] << std::endl;
   }
@@ -48,7 +48,7 @@ int main() {
   z >>= vz;
   std::shared_ptr<TrivialTrace<double>> trace = ReverseAD::trace_off<double>();
   std::cout << "z = " << z.getVal() << std::endl;
-  std::shared_ptr<DerivativeTensor<int, double>> tensor;
+  std::shared_ptr<DerivativeTensor<size_t, double>> tensor;
   BaseReverseThird<double> third(trace);
   tensor = third.compute(4, 1).get_tensor();
   check_answer(tensor);
